Remove editor menu
The primary change of this commit is the removal of the editor menu class.
The counters for the various tiles have never been confirmed in the design
of the game and will probably not be necessary anymore, or at least not
for a long time.

A few components have also been transitioned to GDScirpt:
- The label value item has been completely removed
- The paper panel shader management script is now in GDScript

In further steps, we need to figure out how to best route input from
GDScript to the C# data backend. The architecture there needs to be well
thought through.