Initiating reversal to full C# implementation
As usual with these engines, cross scripting functionality seems to be
only minimally supported. Nothing comes even close to the easy of use of
blueprints in Unreal Engine.

NOTE (MM):
I tried for four hours to make this work and get access to my C# classes
and data in GDScript and it simply won't work. At this point, making the
interop work nicely is a pain in the butt and thus, I refuse to continue
with it. I'll slowly revert everything back to C# and will ignore GDScript
from hereon out most of the time.