Spawn and display three roman units
The whole proces is quite manual at this point. Things are hooked together
very explicitly. It's not entirely clear to me just yet what the architecture
should be in the end. I like what I had originally design with the layers,
maps and the whole architecture around it.

However, how coordinate updates are propagated through the system, is not
clear yet, however, it seems more and more likely that we need a specialized
storage for classes that expose a coordinates property.