Remove locations from entity classes
This is a major design problem for my code base. We want our locations
to be stored in a single place, but we don't want to have high coupling
between our units and the maps they are stored in. Put differently, units
should not need to know where their positions are stored.

This leads to approaches where either we can not store the position of
the unit on the unit, or we need to make sure that all of our index classes
are updated every time the location of a unit changes.

The observer pattern would be one way to go here. Another way to go would
be a reactive pattern, where all of our entity stats are observable values
and would broadcast value changes. This could also be highly benefical
later down the road.

This is where RX.net comes in and could make things quite easy. The downside
to this is that we don't actually know anymore what the overhead of changing
the position of one of our units is. Being explicit about that would help a lot.