Tie entity placement to map context
I don't think this solution is particularly elegant, to be honest. But
it will do for now. The core of it is that entities need to return an
axial coordinate as their location (but they can also not have a location
at all, if they haven't been placed yet).

Assignments to this location however don't need to be supported.
How do we ensure then that movements are kept in sync with the map context?

We do this by creating the placement wrapper directly in the map context.
Only the map context has access to the placemen and units can only be
moved through the map context. There are issues with this design as well,
first among them the fact that it seems weird that we can not change the
position of units anymore. this is far from data oriented design.