Improve the editor menu UI slightly and use toggle states on buttons (#142)
THe previous implementation was a really good start, but instead of
simply having button pressed that don't properly show the active tile
type, I have now switched the buttons to being toggle buttons.

I also adjusted the paper background on the menu page to use the proper
paper material.

I furthermore made adjustments to the theme and used it everywhere. The
theme now also has nicer colors for the buttons.